+++ test/test_netconf.sh
@@ -46,7 +46,7 @@ if [ $BE -ne 0 ]; then
     start_backend -s init -f $cfg -- -s
 
     new "waiting"
-    sleep $RCWAIT
+    wait_backend
 fi
 
 new "netconf hello"