+++ CHANGELOG.md
@@ -237,6 +237,7 @@
 	
 ### Corrected Bugs
 
+* Fixed a problem with some netconf error messages caused restconf daemon to exit due to no XML encoding
 * Check cligen tab mode, dont start if CLICON_CLI_TAB_MODE is undefined
 * Startup transactions did not mark added tree with XML_FLAG_ADD as it should.
 * Restconf PUT different keys detected (thanks @dcornejo) and fixed